Lightweight Sinatra frontend; fix URL validation
This adds a lightweight Sinatra frontend inspired by the
Python-based grandparent of this project: http://gyp.se. The
frontend mostly enables calling the API as an anonymous user for
the purposes of shortening a URL, and implements the actual 301
redirect that URL shorteners are famous for!

URL redirections are now logged. Much more gloss and utility will
come from this front end. On the shortlist: URL visit statistics
and API key management.

This commit also fixes a URL validation error where a user could
submit simply `http(s)://` and have it accepted as a valid URL.
The area of URL validation could get a lot better -- Miny's
predecessor went as far as checking URL blacklists before
shortening or redirecting.

Thanks to Eric Meyer for the CSS reset included.
 - http://meyerweb.com/eric/tools/css/reset/

Also thanks to the wonderful, ubiquitous jQuery library.
 - http://jquery.com